+++ crates/holochain_p2p/src/spawn/actor.rs
@@ -1052,6 +1052,14 @@ impl HolochainP2pActor {
             config.set_module_config(&iroh_transport_config)?;
             override_needed = true;
         }
+        if let Some(auth_material) = space_overrides.base64_auth_material.as_ref() {
+            // get current bootstrap config and override auth_material_base64
+            let mut core_bootstrap_config: kitsune2_core::factories::CoreBootstrapModConfig =
+                config.get_module_config().unwrap_or_default();
+            core_bootstrap_config.core_bootstrap.auth_material_base64 = Some(auth_material.clone());
+            config.set_module_config(&core_bootstrap_config)?;
+            override_needed = true;
+        }
         #[cfg(feature = "transport-tx5-backend-go-pion")]
         if let Some(relay_url) = space_overrides.relay_url.as_ref() {
             // get current tx5 transport config and override server_url
@@ -2696,6 +2704,7 @@ mod tests {
         let space_overrides = CellConfigOverrides {
             bootstrap_url: Some("http://override:1234".to_string()),
             relay_url: Some("wss://override:5678".to_string()),
+            base64_auth_material: None,
         };
         let overrides = actor_p2p
             .space_config_override(space_overrides)
@@ -2730,6 +2739,7 @@ mod tests {
         let space_overrides = CellConfigOverrides {
             bootstrap_url: Some("http://override:1234".to_string()),
             relay_url: Some("wss://override:5678".to_string()),
+            base64_auth_material: None,
         };
         let overrides = actor_p2p
             .space_config_override(space_overrides)