+++ crates/holochain_util/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## \[Unreleased\]
 
+- Added `optional_url2_schema` helper to the `jsonschema` module for JSON Schema generation of `Option<Url2>` fields.
+
 ## 0.7.0-dev.1
 
 ## 0.7.0-dev.0